+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
 
 - Added new flag `hostname` in `safira function deploy`
+- Added config scale in a stack.yaml
 
 ## Version v0.0.2 - 2020-06-08